Extract HTTP assertions in a separate class
