+++ Resources/config/functional_test.xml
@@ -27,5 +27,8 @@
         <service id="liip_functional_test.fixtures_loader" class="Liip\FunctionalTestBundle\Utils\FixturesLoader">
             <argument type="service" id="service_container" />
         </service>
+
+        <service id="liip_functional_test.http_assertions" class="Liip\FunctionalTestBundle\Utils\HttpAssertions">
+        </service>
     </services>
 </container>
